loading data from array into the view
