+++ Turfy/Turfy/RecipientListViewController.swift
@@ -1,0 +1,50 @@
+//
+//  RecipientListViewController.swift
+//  Turfy
+//
+//  Created by James Stonehill on 01/11/2016.
+//  Copyright © 2016 Lawrence Dawson. All rights reserved.
+//
+
+import UIKit
+
+class RecipientListViewController: UIViewController, UITableViewDataSource {
+
+    let data:[String] = ["ONE","TWO","Three"]
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
+        let cell = UITableViewCell()
+        
+        cell.textLabel?.text = data[indexPath.row]
+        
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return data.count
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
