--- Turfy/Turfy/RecipientSearchTableViewController.swift
@@ -1,148 +1,0 @@
-//
-//  RecipientSearchTableViewController.swift
-//  Turfy
-//
-//  Created by Tam Borine on 31/10/2016.
-//  Copyright © 2016 Lawrence Dawson. All rights reserved.
-//
-
-import UIKit
-
-class RecipientSearchTableViewController: UITableViewController, UITableViewDataSource {
-    
-    var dataArray: [String] = ["ONE", "TWO"]
-    var filteredArray = [String]()
-    var shouldShowSearchResults = false
-    
-    @IBOutlet weak var tblSearchResults: UITableView!
-
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        tblSearchResults.delegate = self
-        tblSearchResults.dataSource = self
-        tblSearchResults.reloadData()
-        
-        DispatchQueue.main.async(execute: {self.refreshTable()})
-        
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        // self.navigationItem.rightBarButtonItem = self.editButtonItem()
-        
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-    // MARK: - Table view data source
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 1
-    }
-    
-    
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        
-        /*
-        if shouldShowSearchResults {
-            return filteredArray.count
-        }
-        else {
-            return dataArray.count
-        }
-         */
-        
-        return dataArray.count
-    }
-    
-    /*
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        print("in tableView")
-        
-       // tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-        
-        var cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath as IndexPath) as! RecipientSearchTableViewCell
-        cell.alabel?.text = dataArray[0]
-
-//        if shouldShowSearchResults {
-//            cell.textLabel?.text = filteredArray[indexPath.row]
-//            print("tests if it is in the true part of the statement")
-//        }
-//        else {
-//            print("should be inputing the text into the cell")
-//            print("\(dataArray[0])")
-//            //cell.textLabel?.text = dataArray[indexPath.row]
-//            cell.textLabel?.text = dataArray[0]
-//        }
-        
-        return cell
-    }
-    */
-    
-    func refreshTable(){
-        self.tableView.reloadData()
-    }
-
-
-    
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "reuseIdentifier", for: indexPath)
-
-        // Configure the cell...
-        cell.textLabel?.text = "CEll"
-        
-        return cell
-    }
-    
-
-    /*
-    // Override to support conditional editing of the table view.
-    override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the specified item to be editable.
-        return true
-    }
-    */
-
-    /*
-    // Override to support editing the table view.
-    override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
-    }
-    */
-
-    /*
-    // Override to support rearranging the table view.
-    override func tableView(_ tableView: UITableView, moveRowAt fromIndexPath: IndexPath, to: IndexPath) {
-
-    }
-    */
-
-    /*
-    // Override to support conditional rearranging of the table view.
-    override func tableView(_ tableView: UITableView, canMoveRowAt indexPath: IndexPath) -> Bool {
-        // Return false if you do not want the item to be re-orderable.
-        return true
-    }
-    */
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
